rename WiflyLight to WyLight
