+++ android/WiflyLight/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="biz.bruenn.WiflyLight"
-    android:versionCode="29"
-    android:versionName="0.29" >
+    package="biz.bruenn.WyLight"
+    android:versionCode="30"
+    android:versionName="0.30" >
 
     <uses-sdk
         android:minSdkVersion="15"
