+++ android/WiflyLight/res/values/strings.xml
@@ -4,7 +4,7 @@
     <string name="action_add">Add</string>
     <string name="action_delete">Delete</string>
     <string name="action_stop">Stop</string>
-    <string name="app_name">WiflyLight</string>
+    <string name="app_name">WyLight</string>
     <string name="cancel">Cancel</string>
     <string name="clear">Clear</string>
     <string name="configure">configure</string>
