+++ android/WiflyLight/src/biz/bruenn/WyLight/EditCommandActivity.java
@@ -1,6 +1,6 @@
 package biz.bruenn.WyLight;
 
-import biz.bruenn.WiflyLight.R;
+import biz.bruenn.WyLight.R;
 import biz.bruenn.WyLight.view.RgbVolumeView;
 import android.app.Activity;
 import android.os.Bundle;
