+++ android/WiflyLight/src/biz/bruenn/WyLight/view/RgbVolumeView.java
@@ -1,6 +1,6 @@
 package biz.bruenn.WyLight.view;
 
-import biz.bruenn.WiflyLight.R;
+import biz.bruenn.WyLight.R;
 import biz.bruenn.WyLight.OnColorChangeListener;
 import android.content.Context;
 import android.graphics.Color;
